+++ Shmungus/src/engine/ShmingoCore.h
@@ -19,7 +19,7 @@
 
 //Other macros
 #define se_currentWorld se_application.getCurrentWorld()
-#define se_deltaTime se_application.getDeltaTime()
+#define se_deltaTime (float)se_application.getDeltaTime()
 #define se_timeElapsed se_application.getTimeElapsed()
 #define se_onTick se_application.getOnTick()
